pyproject.toml
[project]
name = "sverigesforetagsguide.se"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyascii==0.3.2",
    "asgiref==3.8.1",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "autopep8==2.3.2",
    "beautifulsoup4==4.13.4",
    "black==25.1.0",
    "cattrs==24.1.3",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "cryptography==45.0.3",
    "cssbeautifier==1.15.4",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "django==5.2.1",
    "django-allauth==65.8.1",
    "django-appconf==1.1.0",
    "django-debug-toolbar==5.2.0",
    "django-extensions==4.1",
    "django-filter==25.1",
    "django-modelcluster==6.4",
    "django-modeltranslation==0.19.14",
    "django-mptt==0.18.0",
    "django-permissionedforms==0.1",
    "django-rosetta==0.10.2",
    "django-select2==8.4.1",
    "django-stubs-ext==5.2.0",
    "django-taggit==6.1.0",
    "django-tasks==0.7.0",
    "django-treebeard==4.7.1",
    "django-webpack-loader==3.2.0",
    "djangorestframework==3.16.0",
    "djlint==1.36.4",
    "docstring-to-markdown==0.17",
    "draftjs-exporter==5.1.0",
    "editorconfig==0.17.0",
    "et-xmlfile==2.0.0",
    "executing==2.2.0",
    "filetype==1.2.0",
    "flake8==7.2.0",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "ipython==9.2.0",
    "ipython-pygments-lexers==1.1.1",
    "isort==6.0.1",
    "jedi==0.19.2",
    "jedi-language-server==0.45.1",
    "jsbeautifier==1.15.4",
    "json5==0.12.0",
    "laces==0.1.2",
    "lsprotocol==2023.0.1",
    "matplotlib-inline==0.1.7",
    "mccabe==0.7.0",
    "mypy-extensions==1.1.0",
    "nodeenv==1.9.1",
    "oauthlib==3.2.2",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "parso==0.8.4",
    "pathspec==0.12.1",
    "pexpect==4.9.0",
    "pillow==11.2.1",
    "pillow-heif==0.21.0",
    "platformdirs==4.3.8",
    "polib==1.2.0",
    "prompt-toolkit==3.0.51",
    "psycopg==3.2.9",
    "psycopg2==2.9.10",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pycodestyle==2.13.0",
    "pycparser==2.22",
    "pyflakes==3.3.2",
    "pygls==1.3.1",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pyright==1.1.401",
    "python-dotenv==1.1.0",
    "python-slugify==8.0.4",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "six==1.17.0",
    "soupsieve==2.7",
    "sqlparse==0.5.3",
    "stack-data==0.6.3",
    "telepath==0.3.1",
    "text-unidecode==1.3",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typing-extensions==4.13.2",
    "urllib3==2.4.0",
    "wagtail==7.0",
    "wagtail-modeladmin==2.2.0",
    "wagtail-modeltranslation==0.15.2",
    "wcwidth==0.2.13",
    "willow==1.10.0",
    "zipp==3.22.0",
]
